serf: update cmds and change packet size type to u16
It looks like we will need to flush after usb receive,
because the host stalls if nothing is sent after a receive.

+++ fw/system/lib/libcc/src/include/cc/nmac.h
@@ -6,7 +6,7 @@
 #define MAC_PKT_OVERHEAD    7
 #define MAC_PKT_SIZE_MAX    (PHY_FRAME_SIZE_MAX - MAC_PKT_OVERHEAD)
 
-typedef void (* mac_recv_t)(u16 node, u16 peer, u16 dest, size_t size, u8 data[]);
+typedef void (* mac_recv_t)(u16 node, u16 peer, u16 dest, u16 size, u8 data[]);
 
 typedef enum __packed {
     NMAC_SEND_DGRM,
@@ -18,4 +18,4 @@ typedef enum __packed {
 
 bool nmac_init(u16 addr, bool sync_master, mac_recv_t rx);
 u16 nmac_get_addr(void);
-bool nmac_send(nmac_send_t type, u16 dest, u8 size, u8 data[]);
+bool nmac_send(nmac_send_t type, u16 dest, u16 size, u8 data[]);
